test commits for residuals, parsers, trellis and test code
