--- parsers/esm_dictionaries.py
@@ -1,213 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-# vim: tabstop=4 shiftwidth=4 softtabstop=4
-#
-# Copyright (C) 2014-2017 GEM Foundation and G. Weatherill
-#
-# OpenQuake is free software: you can redistribute it and/or modify it
-# under the terms of the GNU Affero General Public License as published
-# by the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# OpenQuake is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Affero General Public License for more details.
-#
-# You should have received a copy of the GNU Affero General Public License
-# along with OpenQuake. If not, see <http://www.gnu.org/licenses/>.
-
-HOUSING = {
-"ACQ": "Acqueducti",
-"BOX": "Box",
-"BRI": "Bridge",
-"BUB": "Building basement",
-"BUI": "Building",
-"CAB": "ENEL Box",
-"CAV": "Cave",
-"DAM": "Dam",
-"EMB": "Embankment",
-"ERR": "Unknown",
-"FIB": "Fiberglass box",
-"GAL": "Tunnel",
-"HIS": "Historical building",
-"HOU": "Small mansory building",
-"POW": "Power plant",
-"QUA": "Quarry",
-"VAU": "Vault",
-"WEL": "Borehole"
-}
-
-
-INSTALLATION = {
-"BF": "Building floor",
-"P": "Pillar",
-"PS": "Building basement",
-"T": "Directly on the ground",
-}
-
-
-INSTRUMENT_TYPE = {
-"A": "Analog",
-"D": "Digital",
-"U": "Unknown"}
-
-PROXIMITY = {
-"0": "Free-field",
-"1": "Close to structure",
-"2": "No information",
-"3": "Inside structure",
-"4": "Structure related free-field"
-}
-
-PROCESSED_CODE = {
-"0": "unprocessed",
-"1": "manually processed",
-"2": "automatically processed",
-"9": "bad quality record",
-}
-
-NETWORK = {
-"3H": "None",
-"4A": "Emersito Seismic Network for Site Effect Studies in L'Aquila town (Central Italy)",
-"4C": "NERA-JRA1-A",
-"4F": "North Texas Earthquake Study: Venus (Johnson County), TX",
-"9A": "East Texas Earthquake Monitoring",
-"A": "Generic Asian Strong Motion Network",
-"AC": "Seismological Network of Albania",
-"AG": "Arkansas Seismic Network",
-"AK": "Alaska Regional Network",
-"AO": "Arkansas Seismic Observatory",
-"AV": "Alaska Volcano Observatory",
-"AY": "Haitian Seismic Network",
-"AZ": "ANZA Regional Network",
-"BA": "University of Basilicata (UNIBAS) Network",
-"BC": "Red Sismica del Noroeste de Mexico",
-"BK": "Berkeley Digital Seismograph Network",
-"BO": "Bosai-Ken Network",
-"BS": "National Seismic Network of Bulgaria",
-"BW": "BayernNetz",
-"C": "Chilean National Seismic Network",
-"C1": "Red Sismologica Nacional",
-"CE": "California Strong Motion Instrumentation Program",
-"CH": "CH Seismic Network",
-"CI": "Southern California Seismic Network",
-"CL": "Corinth Riff Laboratory",
-"CN": "Canadian National Seismic Network",
-"CO": "South Carolina Seismic Network",
-"CQ": "Cyprus Broadband Seismological Network",
-"CR": "Croatian Seismograph Network",
-"CU": "Caribbean Network",
-"CX": "Plate Boundary Observatory Network Northern Chile",
-"CY": "Cayman Islands",
-"CZ": "Czech Regional Seismic Network",
-"DR": "Dominican Republic Seismic Network",
-"E": "ENEA network",
-"EC": "Ecuador Seismic Network",
-"EP": "UTEP Seismic Network",
-"ES": "Spanish Digital Seismic Network",
-"ET": "CERI Southern Appalachian Seismic Network",
-"EU": "Generic European Strong Motion Network",
-"FA": "UCLA Seismic Network",
-"FC": "Generic African Strong Motion Network",
-"FR": "French Broadband Seismological Network, ReNaSS Strasbourg",
-"G": "GEOSCOPE, Institut de Physique du Globe de Paris (IPGP)",
-"GE": "GEOFON",
-"GI": "Red Sismologica Nacional- Guatemala",
-"GM": "None",
-"GR": "German Regional Seismic Network",
-"GS": "US Geological Survey Networks",
-"GU": "Regional Seismic Network of North-western Italy (RSNI)",
-"HI": "ITSAK Strong Motion Network",
-"HK": "Hong Kong Seismograph Network",
-"HL": "National Observatory of Athens Digital Broadband Network",
-"HP": "University of Patras, Seismological Laboratory",
-"HT": "Aristotle University of Thessaloniki Seismological Network",
-"HV": "Hawaiian Volcano Observatory Network",
-"I1": "Iranian Strong Motion Network",
-"IC": "New China Digital Seismograph Network",
-"IP": "Instituto Superior Tecnico Broadband Seismic Network",
-"IS": "Israel National Seismic Network",
-"IT": "Italian Strong Motion Network",
-"IU": "Global Seismograph Network (GSN - IRIS/USGS)",
-"IV": "Italian National Seismic Network",
-"IW": "Intermountain West Seismic Network",
-"IX": "ISNet - Irpinia Seismic Network",
-"JM": "Jamaica Seismograph Network",
-"KO": "Bogazici University Kandilli Observatory And Earthquake Research Institute",
-"KY": "Kentucky Seismic and Strong Motion Network",
-"LC": "LSC (Laboratorio Subterraneo Canfranc)",
-"LE": "Landeserdbebendienst Baden-Wuerttemberg",
-"M": "Generic American Strong Motion Network",
-"MA": "Macedonia Seismological Network (SORM)",
-"MB": "Montana Regional Seismic Network",
-"MD": "Moldova Digital Seismic Network",
-"ME": "Montenegrian Seismic Network",
-"MG": "Seismic Network of the NorthEastern Mexico",
-"MM": "Myanmar National Seismic Network",
-"MN": "Mediterranean Very Broadband Seismographic Network (MNDC)",
-"MS": "Singapore Seismological Network",
-"MT": "Observatory for Multidisciplinary monitoring of Instability of Versant",
-"MX": "Mexican National Seismic Network",
-"N4": "Central and Eastern US Network",
-"NC": "USGS Northern California Regional Network",
-"ND": "New CaleDonia Broadband Seismic Network",
-"NE": "New England Seismic Network",
-"NI": "North-East Italy Broadband Network",
-"NM": "Cooperative New Madrid Seismic Network",
-"NN": "Western Great Basin/Eastern Sierra Nevada",
-"NP": "United States National Strong-Motion Network",
-"NQ": "NetQuakes",
-"NS": "Norwegian National Seismic Network",
-"NU": "Nicaraguan Seismic Network",
-"NV": "Neptune Canada",
-"NZ": "New Zealand National Seismograph Network",
-"OE": "Austrian Seismic Network",
-"OO": "Ocean Observatories Initiative",
-"OV": "Observatorio Vulcanologico y Sismologico de Costa Rica",
-"OX": "North-East Italy Seismic Network",
-"PA": "ChiriNet",
-"PB": "Plate Boundary Observatory Borehole Seismic Network",
-"PG": "Central Coast Seismic Network, PG&E",
-"PR": "Puerto Rico Seismic Network (PRSN) & Puerto Rico Strong Motion Program (PRSMP)",
-"PT": "Pacific Tsunami Warning Seismic System",
-"PY": "Pinyon Flats Observatory (PFO) Array",
-"RA": "Reseau Accelerometrique Permanent (French Accelerometrique Network)",
-"RF": "Friuli Venezia Giulia Accelerommetric Network (RAF)",
-"RM": "Regional Integrated Multi-hazard Early Warning System (RIMES)",
-"RO": "Romanian Seismic Network",
-"S": "Seismographs in Schools Network",
-"SB": "UC Santa Barbara Engineering Seismology Network",
-"SI": "Province Sudtirol",
-"SL": "Slovenia Seismic Network",
-"SN": "Southern Great Basin Network",
-"SS": "Single Station: generic network code for any network in US",
-"ST": "Trentino Seismic Network",
-"SV": "Servicio Nacional de Estudios Territoriales (SNET), El Salvador",
-"TA": "USArray Transportable Array",
-"TH": "Thuringer Seismisches Netz (TSN)",
-"TK": "National Strong-Motion Network of Turkey (TR-NSMN)",
-"TS": "TERRAscope (Southern California Seismic Network)",
-"TV": "INGV Experiments Network",
-"TX": "Texas Seismological Network",
-"US": "United States National Seismic Network",
-"UW": "Pacific Northwest Regional Seismic Network",
-"UZ": "Uzbekistan Digital Seismic Network",
-"WI": "West Indies IPGP Network",
-"WM": "Western Mediterranean Seismic Network",
-"WR": "California Division of Water Resources",
-"XO": "Seismic Network for Site Effect Studies in Amatrice Area (Central Italy)",
-"XS": "Maule Earthquake (Chile) Aftershock Experiment",
-"XY": "Chile RAMP",
-"Y3": "Mogul Aftershocks",
-"YB": "Haiti Earthquake Aftershock Experiment",
-"YD": "Calabria-Appennine-Tyrrhenian/Subduction-Collision-Accretion Network",
-"YI": "Retreating-Trench, Extension, and Accretion LTectonics: A Multidisciplinary Study of the Northern Apennines",
-"YN": "San Jacinto Fault Zone",
-"YP": "Argostoli earthquake aftershock experiment",
-"YX": "2008 SoSAF Ramp Test",
-"Z3": "AlpArray",
-"ZN": "project SISMOVALP temporary network",
-"ZW": "North Texas Earthquake Study: Azle and Irving/Dallas",
-}
-
